Found issue with doors
I think I figured the issue with the doors.
The mesh doesn't rebake in realtime.
I installed a module to unity to let me have mesh components I can put on the map, and I can later have these components rebake the mesh everytime an item is placed on it.
It's not exactly it yet, but I am getting much closer.